--- docs/api/tmtccmd.ccsds.rst
@@ -1,21 +1,0 @@
-tmtccmd.ccsds package
-=====================
-
-Submodules
-----------
-
-tmtccmd.ccsds.handler module
---------------------------------
-
-.. automodule:: tmtccmd.ccsds.handler
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: tmtccmd.ccsds
-   :members:
-   :undoc-members:
-   :show-inheritance:
