+++ docs/api/tmtccmd.core.rst
@@ -1,21 +1,29 @@
 tmtccmd.core package
 ====================
 
-Submodules
-----------
+Module contents
+---------------
 
-tmtccmd.core.frontend module
-----------------------------
-
-.. automodule:: tmtccmd.core.frontend
+.. automodule:: tmtccmd.core
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.core.frontend\_base module
+Submodules
+----------
+
+tmtccmd.core.ccsds\_backend module
 ----------------------------------
 
-.. automodule:: tmtccmd.core.frontend_base
+.. automodule:: tmtccmd.core.ccsds_backend
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tmtccmd.core.base module
+----------------------------
+
+.. automodule:: tmtccmd.core.base
    :members:
    :undoc-members:
    :show-inheritance:
@@ -24,14 +32,6 @@ tmtccmd.core.globals\_manager module
 ------------------------------------
 
 .. automodule:: tmtccmd.core.globals_manager
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: tmtccmd.core
    :members:
    :undoc-members:
    :show-inheritance:
